refactor(color-text): move color parsing logic to utils
Move color text parsing functionality from color-text.tsx to utils.ts to improve code organization and reusability. Also add 'analyze' and 'cache' to gitignore.
