+++ .gitignore
@@ -44,3 +44,6 @@ next-env.d.ts
 # Sentry Config File
 .env.sentry-build-plugin
 .env
+
+analyze
+cache
